agent/requirements.txt
aiohttp==3.10.6
python-dotenv==1.0.1
